refactor: split map feature logic into focused utilities
