+++ src/areacode/features/map/components/MapLayers.tsx
@@ -11,23 +11,7 @@ import {
   maFillStyle,
   maLabelStyle,
 } from '../mapStyles'
-
-function flattenCoordinates(geometry: Geometry): Position[] {
-  switch (geometry.type) {
-    case 'Point':
-      return [geometry.coordinates]
-    case 'MultiPoint':
-    case 'LineString':
-      return geometry.coordinates
-    case 'MultiLineString':
-    case 'Polygon':
-      return geometry.coordinates.flat(1)
-    case 'MultiPolygon':
-      return geometry.coordinates.flat(2)
-    default:
-      return []
-  }
-}
+import { getLabelPosition } from '../utils/geometry'
 
 function getDigits2FontSize(zoom: number): number {
   const minZoom = 5
@@ -95,27 +79,6 @@ function getAdjustedPrefPosition(
   }
 
   return basePosition
-}
-
-function getLabelPosition(geometry: Geometry): Position | null {
-  const points = flattenCoordinates(geometry)
-  if (points.length === 0) {
-    return null
-  }
-
-  let minLng = Infinity
-  let maxLng = -Infinity
-  let minLat = Infinity
-  let maxLat = -Infinity
-
-  points.forEach(([lng, lat]) => {
-    if (lng < minLng) minLng = lng
-    if (lng > maxLng) maxLng = lng
-    if (lat < minLat) minLat = lat
-    if (lat > maxLat) maxLat = lat
-  })
-
-  return [(minLng + maxLng) / 2, (minLat + maxLat) / 2]
 }
 
 export function MapLayers({
